binding generation: support non-gpu builds of skia-bindings
