+++ CMakeLists.txt
@@ -9,6 +9,8 @@ add_compile_definitions(SK_SHAPER_HARFBUZZ_AVAILABLE, SK_VULKAN, SK_XML)
 
 add_library(skiabindings
         skia-bindings/src/bindings.cpp
+        skia-bindings/src/gl.cpp
+        skia-bindings/src/gpu.cpp
         skia-bindings/src/paragraph.cpp
         skia-bindings/src/shaper.cpp 
         skia-bindings/src/svg.cpp
