+++ skia-bindings/src/bindings.h
@@ -10,6 +10,11 @@ inline sk_sp<T> spFromConst(const T* pt) {
     return sk_sp<T>(const_cast<T*>(pt));
 }
 
+template<typename T>
+inline sk_sp<T> sp(T* pt) {
+    return sk_sp<T>(pt);
+}
+
 // Used in textlayout::Paragraph::findTypefaces()
 
 struct SkStrings {
